refactor(configure): extract entity loader
Configuration files are Velocity templates, so their variables must be
expanded before the content is deserialized. That logic was private to
AbstractImporter and thus only available to the importers.

Move it into a ConfigurationEntityLoader, so that everything reading
configuration files applies the same expansion. Without it, a comparison
against the server would report the unexpanded '$env.KCC_*' placeholders
as differences.

AbstractImporter delegates to it and keeps its loadEntity signatures, so
none of the importers change.
